consolidate initializers
