+++ services/horizon/internal/init.go
@@ -1,72 +1,214 @@
 package horizon
 
 import (
+	"net/http"
+	"net/url"
+	"os"
+	"time"
+
+	raven "github.com/getsentry/raven-go"
+	"github.com/gomodule/redigo/redis"
+	metrics "github.com/rcrowley/go-metrics"
+	"github.com/stellar/go/services/horizon/internal/db2/core"
+	"github.com/stellar/go/services/horizon/internal/db2/history"
+	"github.com/stellar/go/services/horizon/internal/ingest"
+	"github.com/stellar/go/services/horizon/internal/txsub"
+	results "github.com/stellar/go/services/horizon/internal/txsub/results/db"
+	"github.com/stellar/go/services/horizon/internal/txsub/sequence"
+	"github.com/stellar/go/support/db"
 	"github.com/stellar/go/support/log"
 )
 
-// InitFn is a function that contributes to the initialization of an App struct
-type InitFn func(*App)
-
-type initializer struct {
-	Name string
-	Fn   InitFn
-	Deps []string
-}
-
-type initializerSet []initializer
-
-var appInit initializerSet
-
-// Add adds a new initializer into the chain
-func (is *initializerSet) Add(name string, fn InitFn, deps ...string) {
-	*is = append(*is, initializer{
-		Name: name,
-		Fn:   fn,
-		Deps: deps,
-	})
-}
-
-// Run initializes the provided application, but running every Initializer
-func (is *initializerSet) Run(app *App) {
-	init := *is
-	alreadyRun := make(map[string]bool)
-
-	for {
-		ranInitializer := false
-		for _, i := range init {
-			runnable := true
-
-			// if we've already been run, skip
-			if _, ok := alreadyRun[i.Name]; ok {
-				runnable = false
-			}
-
-			// if any of our dependencies haven't been run, skip
-			for _, d := range i.Deps {
-				if _, ok := alreadyRun[d]; !ok {
-					runnable = false
-					break
-				}
-			}
-
-			if !runnable {
-				continue
-			}
-
-			log.WithField("init_name", i.Name).Debug("running initializer")
-			i.Fn(app)
-			alreadyRun[i.Name] = true
-			ranInitializer = true
-		}
-		// If, after a full loop through the initializers we ran nothing
-		// we are done
-		if !ranInitializer {
-			break
-		}
+func initHorizonDb(app *App) {
+	session, err := db.Open("postgres", app.config.DatabaseURL)
+	if err != nil {
+		log.Panic(err)
 	}
 
-	// if we didn't get to run all initializers, we have a cycle
-	if len(alreadyRun) != len(init) {
-		log.Panic("initializer cycle detected")
+	// Make sure MaxIdleConns is equal MaxOpenConns. In case of high variance
+	// in number of requests closing and opening connections may slow down Horizon.
+	session.DB.SetMaxIdleConns(app.config.MaxDBConnections)
+	session.DB.SetMaxOpenConns(app.config.MaxDBConnections)
+	app.historyQ = &history.Q{session}
+}
+
+func initCoreDb(app *App) {
+	session, err := db.Open("postgres", app.config.StellarCoreDatabaseURL)
+	if err != nil {
+		log.Panic(err)
+	}
+
+	// Make sure MaxIdleConns is equal MaxOpenConns. In case of high variance
+	// in number of requests closing and opening connections may slow down Horizon.
+	session.DB.SetMaxIdleConns(app.config.MaxDBConnections)
+	session.DB.SetMaxOpenConns(app.config.MaxDBConnections)
+	app.coreQ = &core.Q{session}
+}
+
+func initIngester(app *App) {
+	if !app.config.Ingest {
+		return
+	}
+
+	if app.config.NetworkPassphrase == "" {
+		log.Error("Cannot start ingestion without network passphrase. Please confirm connectivity with stellar-core.")
+		os.Exit(1)
+	}
+
+	app.ingester = ingest.New(
+		app.config.NetworkPassphrase,
+		app.config.StellarCoreURL,
+		app.CoreSession(nil),
+		app.HorizonSession(nil),
+		ingest.Config{
+			EnableAssetStats:         app.config.EnableAssetStats,
+			IngestFailedTransactions: app.config.IngestFailedTransactions,
+		},
+	)
+
+	app.ingester.SkipCursorUpdate = app.config.SkipCursorUpdate
+	app.ingester.HistoryRetentionCount = app.config.HistoryRetentionCount
+}
+
+// initSentry initialized the default sentry client with the configured DSN
+func initSentry(app *App) {
+	if app.config.SentryDSN == "" {
+		return
+	}
+
+	log.WithField("dsn", app.config.SentryDSN).Info("Initializing sentry")
+	err := raven.SetDSN(app.config.SentryDSN)
+	if err != nil {
+		panic(err)
+	}
+}
+
+// initLogglyLog attaches a loggly hook to our logging system.
+func initLogglyLog(app *App) {
+	if app.config.LogglyToken == "" {
+		return
+	}
+
+	log.WithFields(log.F{
+		"token": app.config.LogglyToken,
+		"tag":   app.config.LogglyTag,
+	}).Info("Initializing loggly hook")
+
+	hook := log.NewLogglyHook(app.config.LogglyToken, app.config.LogglyTag)
+	log.DefaultLogger.Logger.Hooks.Add(hook)
+
+	go func() {
+		<-app.ctx.Done()
+		hook.Flush()
+	}()
+}
+
+func initDbMetrics(app *App) {
+	app.historyLatestLedgerGauge = metrics.NewGauge()
+	app.historyElderLedgerGauge = metrics.NewGauge()
+	app.coreLatestLedgerGauge = metrics.NewGauge()
+	app.horizonConnGauge = metrics.NewGauge()
+	app.coreConnGauge = metrics.NewGauge()
+	app.goroutineGauge = metrics.NewGauge()
+	app.metrics.Register("history.latest_ledger", app.historyLatestLedgerGauge)
+	app.metrics.Register("history.elder_ledger", app.historyElderLedgerGauge)
+	app.metrics.Register("stellar_core.latest_ledger", app.coreLatestLedgerGauge)
+	app.metrics.Register("history.open_connections", app.horizonConnGauge)
+	app.metrics.Register("stellar_core.open_connections", app.coreConnGauge)
+	app.metrics.Register("goroutines", app.goroutineGauge)
+}
+
+func initIngesterMetrics(app *App) {
+	if app.ingester == nil {
+		return
+	}
+	app.metrics.Register("ingester.ingest_ledger",
+		app.ingester.Metrics.IngestLedgerTimer)
+	app.metrics.Register("ingester.clear_ledger",
+		app.ingester.Metrics.ClearLedgerTimer)
+}
+
+func initTxSubMetrics(app *App) {
+	app.submitter.Init()
+	app.metrics.Register("txsub.buffered", app.submitter.Metrics.BufferedSubmissionsGauge)
+	app.metrics.Register("txsub.open", app.submitter.Metrics.OpenSubmissionsGauge)
+	app.metrics.Register("txsub.succeeded", app.submitter.Metrics.SuccessfulSubmissionsMeter)
+	app.metrics.Register("txsub.failed", app.submitter.Metrics.FailedSubmissionsMeter)
+	app.metrics.Register("txsub.total", app.submitter.Metrics.SubmissionTimer)
+}
+
+// initWebMetrics registers the metrics for the web server into the provided
+// app's metrics registry.
+func initWebMetrics(app *App) {
+	app.metrics.Register("requests.total", app.web.requestTimer)
+	app.metrics.Register("requests.succeeded", app.web.successMeter)
+	app.metrics.Register("requests.failed", app.web.failureMeter)
+}
+
+func initRedis(app *App) {
+	if app.config.RedisURL == "" {
+		return
+	}
+
+	redisURL, err := url.Parse(app.config.RedisURL)
+	if err != nil {
+		log.Panic(err)
+	}
+
+	app.redis = &redis.Pool{
+		MaxIdle:     3,
+		IdleTimeout: 240 * time.Second,
+		Dial:        dialRedis(redisURL),
+		TestOnBorrow: func(c redis.Conn, t time.Time) error {
+			_, err := c.Do("PING")
+			return err
+		},
+	}
+
+	// test the connection
+	c := app.redis.Get()
+	defer c.Close()
+
+	_, err = c.Do("PING")
+	if err != nil {
+		log.Panic(err)
+	}
+}
+
+func dialRedis(redisURL *url.URL) func() (redis.Conn, error) {
+	return func() (redis.Conn, error) {
+		c, err := redis.Dial("tcp", redisURL.Host)
+		if err != nil {
+			return nil, err
+		}
+
+		if redisURL.User == nil {
+			return c, err
+		}
+
+		if pass, ok := redisURL.User.Password(); ok {
+			if _, err = c.Do("AUTH", pass); err != nil {
+				c.Close()
+				return nil, err
+			}
+		}
+
+		return c, err
+	}
+}
+
+func initSubmissionSystem(app *App) {
+	cq := &core.Q{Session: app.CoreSession(nil)}
+
+	app.submitter = &txsub.System{
+		Pending:         txsub.NewDefaultSubmissionList(),
+		Submitter:       txsub.NewDefaultSubmitter(http.DefaultClient, app.config.StellarCoreURL),
+		SubmissionQueue: sequence.NewManager(),
+		Results: &results.DB{
+			Core:    cq,
+			History: &history.Q{Session: app.HorizonSession(nil)},
+		},
+		Sequences:         cq.SequenceProvider(),
+		NetworkPassphrase: app.config.NetworkPassphrase,
 	}
 }
