--- services/horizon/internal/init_core_info.go
@@ -1,9 +1,0 @@
-package horizon
-
-func initStellarCoreInfo(app *App) {
-	app.UpdateStellarCoreInfo()
-}
-
-func init() {
-	appInit.Add("stellarCoreInfo", initStellarCoreInfo, "app-context", "log")
-}
