--- services/horizon/internal/init_path_finder.go
@@ -1,13 +1,0 @@
-package horizon
-
-import (
-	"github.com/stellar/go/services/horizon/internal/simplepath"
-)
-
-func initPathFinding(app *App) {
-	app.paths = &simplepath.Finder{app.CoreQ()}
-}
-
-func init() {
-	appInit.Add("path-finder", initPathFinding, "app-context", "log", "core-db")
-}
