--- services/horizon/internal/init_txsub.go
@@ -1,31 +1,0 @@
-package horizon
-
-import (
-	"net/http"
-
-	"github.com/stellar/go/services/horizon/internal/db2/core"
-	"github.com/stellar/go/services/horizon/internal/db2/history"
-	"github.com/stellar/go/services/horizon/internal/txsub"
-	results "github.com/stellar/go/services/horizon/internal/txsub/results/db"
-	"github.com/stellar/go/services/horizon/internal/txsub/sequence"
-)
-
-func initSubmissionSystem(app *App) {
-	cq := &core.Q{Session: app.CoreSession(nil)}
-
-	app.submitter = &txsub.System{
-		Pending:         txsub.NewDefaultSubmissionList(),
-		Submitter:       txsub.NewDefaultSubmitter(http.DefaultClient, app.config.StellarCoreURL),
-		SubmissionQueue: sequence.NewManager(),
-		Results: &results.DB{
-			Core:    cq,
-			History: &history.Q{Session: app.HorizonSession(nil)},
-		},
-		Sequences:         cq.SequenceProvider(),
-		NetworkPassphrase: app.config.NetworkPassphrase,
-	}
-}
-
-func init() {
-	appInit.Add("txsub", initSubmissionSystem, "app-context", "log", "horizon-db", "core-db")
-}
