+++ services/horizon/internal/web.go
@@ -206,28 +206,3 @@ func remoteAddrIP(r *http.Request) string {
 func firstXForwardedFor(r *http.Request) string {
 	return strings.TrimSpace(strings.SplitN(r.Header.Get("X-Forwarded-For"), ",", 2)[0])
 }
-
-func init() {
-	appInit.Add(
-		"web.init",
-		initWeb,
-		"app-context",
-	)
-
-	appInit.Add(
-		"web.rate-limiter",
-		initWebRateLimiter,
-		"web.init",
-	)
-	appInit.Add(
-		"web.middleware",
-		initWebMiddleware,
-		"web.init",
-		"web.rate-limiter",
-	)
-	appInit.Add(
-		"web.actions",
-		initWebActions,
-		"web.init",
-	)
-}
